ai-engine/requirements.txt
fastapi
uvicorn
python
python-multipart
deepface
tf-keras
retina-face
pytesseract
pillow
numpy
